Feature: Static info page, rearranged helpers, created and deleted dupe route
